Download a file from a GitHub repository

This function downloads a file from a specified GitHub repository and saves it to a local directory.

Usage

download_file(file_path, dest_path, verbose = FALSE)

Arguments

file_path

A character string specifying the path to the file in the GitHub repository.

dest_path

A character string specifying the destination directory where the file will be saved.

verbose

A logical value indicating whether to print messages during the download process. Default is FALSE.

Value

A character string representing the path to the downloaded file, or NULL if the download fails.